Finally got bloom effect to an appropriate state

+++ Doom/src/Doom/Render/TextureAtlas.cpp
@@ -20,14 +20,14 @@ Doom::TextureAtlas * Doom::TextureAtlas::CreateTextureAtlas(std::string name, fl
 
 float * Doom::TextureAtlas::GetSpriteUVs(float x, float y)
 {
-	m_UVs[0] = (x * m_SpriteWidth) / m_Texture->m_width;
-	m_UVs[1] = (y * m_SpriteHeight) / m_Texture->m_height;
-	m_UVs[2] = ((x + 1) * m_SpriteWidth) / m_Texture->m_width;
-	m_UVs[3] = (y * m_SpriteHeight) / m_Texture->m_height;
-	m_UVs[4] = ((x + 1) * m_SpriteWidth) / m_Texture->m_width;
-	m_UVs[5] = ((y + 1) * m_SpriteHeight) / m_Texture->m_height;
-	m_UVs[6] = (x * m_SpriteWidth) / m_Texture->m_width;
-	m_UVs[7] = ((y + 1) * m_SpriteHeight) / m_Texture->m_height;
+	m_UVs[0] = (x * m_SpriteWidth) / (float)m_Texture->m_width;
+	m_UVs[1] = (y * m_SpriteHeight) / (float)m_Texture->m_height;
+	m_UVs[2] = ((x + 1) * m_SpriteWidth) / (float)m_Texture->m_width;
+	m_UVs[3] = (y * m_SpriteHeight) / (float)m_Texture->m_height;
+	m_UVs[4] = ((x + 1) * m_SpriteWidth) / (float)m_Texture->m_width;
+	m_UVs[5] = ((y + 1) * m_SpriteHeight) / (float)m_Texture->m_height;
+	m_UVs[6] = (x * m_SpriteWidth) / (float)m_Texture->m_width;
+	m_UVs[7] = ((y + 1) * m_SpriteHeight) / (float)m_Texture->m_height;
 	return m_UVs;
 }
